+++ include/EclipseMonitor/Trie/BranchNode.hpp
@@ -5,118 +5,124 @@
 
 #pragma once
 
-#include "EclipseMonitor/EthKeccak256.hpp"
-#include "EclipseMonitor/Trie/Nibbles.hpp"
-#include "EclipseMonitor/Trie/TrieNode.hpp"
+#include "../EthKeccak256.hpp"
+#include "../Trie/Nibbles.hpp"
+#include "../Trie/TrieNode.hpp"
 #include "SimpleObjects/Internal/make_unique.hpp"
-
-using namespace SimpleObjects;
 
 namespace EclipseMonitor
 {
 
-namespace Branch
+namespace Trie
 {
 
-class BranchNode : public TrieNode::NodeBase
+class BranchNode : public NodeBase
 {
 public:
 
-    BranchNode()
-    {
-        NodeHasValue = false;
-        Branches.resize(16);
-    }
+	BranchNode()
+	{
+		NodeHasValue = false;
+		Branches.resize(16);
+	}
 
-    bool HasValue()
-    {
-        return NodeHasValue;
-    }
+	bool HasValue()
+	{
+		return NodeHasValue;
+	}
 
-    void SetBranch(const Nibble& nibble,
-                   std::unique_ptr<TrieNode::NodeBase> other)
-    {
-        int nibbleInt = static_cast<int>(nibble);
-        Branches[nibbleInt] = std::move(other);
-    }
+	void SetBranch(const Nibble& nibble,
+				   std::unique_ptr<Node> other)
+	{
+		int nibbleInt = static_cast<int>(nibble);
+		Branches[nibbleInt] = std::move(other);
+	}
 
-    void RemoveBranch(const Nibble& nibble)
-    {
-        int nibbleInt = static_cast<int>(nibble);
-        Branches[nibbleInt].reset();
-    }
+	std::unique_ptr<Node>& GetBranch(Nibble nibble)
+	{
+		int nibbleInt = static_cast<int>(nibble);
+		return Branches[nibbleInt];
+	}
 
-    void SetValue(SimpleObjects::Bytes otherValue)
-    {
-        NodeHasValue = true;
-        Value = std::move(otherValue);
-    }
+	void RemoveBranch(const Nibble& nibble)
+	{
+		int nibbleInt = static_cast<int>(nibble);
+		Branches[nibbleInt].reset();
+	}
 
-    void RemoveValue()
-    {
-        NodeHasValue = false;
-        Value.resize(0);
+	void SetValue(SimpleObjects::Bytes otherValue)
+	{
+		NodeHasValue = true;
+		Value = std::move(otherValue);
+	}
 
-    }
+	void RemoveValue()
+	{
+		NodeHasValue = false;
+		Value.resize(0);
 
-    std::vector<uint8_t> Serialize()
-    {
-        TrieNode::NodeBase* BasePtr = static_cast<TrieNode::NodeBase*>(this);
-        return TrieNode::Node::Serialize(BasePtr);
-    }
+	}
 
-    virtual SimpleObjects::Bytes Hash() override
-    {
-        std::vector<uint8_t> serialized = Serialize();
-        std::array<uint8_t, 32> hashed = EthKeccak256(serialized);
 
-        return SimpleObjects::Bytes(hashed.begin(), hashed.end());
-    }
+	std::vector <uint8_t> Serialize()
+	{
+		NodeBase* BasePtr = static_cast<NodeBase*>(this);
+		return NodeHelper::Serialize(BasePtr);
+	}
 
-    virtual SimpleObjects::List Raw() override
-    {
-        SimpleObjects::List hashes;
-        hashes.resize(NumNodes + 1);
+	virtual NodeType GetNodeType() const override
+	{
+		return NodeType::Branch;
+	}
 
-        for (uint8_t i = 0; i < NumNodes; i++)
-        {
-            if (!Branches[i])
-            {
-                hashes[i] = TrieNode::EmptyNode::EmptyNodeRaw();
-            }
-            else
-            {
-                TrieNode::NodeBase* NodePtr = Branches[i].get();
-                std::vector<uint8_t> serialized = TrieNode::Node::Serialize(NodePtr);
+	virtual SimpleObjects::Bytes Hash() override
+	{
+		std::vector<uint8_t> serialized = Serialize();
+		std::array<uint8_t, 32> hashed = EthKeccak256(serialized);
 
-                if (serialized.size() >= 32)
-                {
-                    hashes[i] = NodePtr->Hash();
-                }
-                else
-                {
-                    hashes[i] = NodePtr->Raw();
-                }
-            }
-        }
+		return SimpleObjects::Bytes(hashed.begin(), hashed.end());
+	}
 
-        hashes[NumNodes] = Value;
-        return hashes;
-    }
+	virtual SimpleObjects::List Raw() override
+	{
+		SimpleObjects::List hashes;
+		hashes.resize(NumNodes + 1);
 
-    std::unique_ptr<TrieNode::NodeBase>& GetBranch(Nibble nibble)
-    {
-        int nibbleInt = static_cast<int>(nibble);
-        return Branches[nibbleInt];
-    }
+		for (uint8_t i = 0; i < NumNodes; i++)
+		{
+			if (!Branches[i])
+			{
+				hashes[i] = EmptyNode::EmptyNodeRaw();
+			}
+			else
+			{
+				// Node* NodePtr = Branches[i].get();
+				std::unique_ptr<NodeBase>& NodeBasePtr = Branches[i]->GetNodeBase();
+				NodeBase* NodePtr = NodeBasePtr.get();
+				std::vector<uint8_t> serialized = NodeHelper::Serialize(NodePtr);
+
+				if (serialized.size() >= 32)
+				{
+					hashes[i] = NodePtr->Hash();
+				}
+				else
+				{
+					hashes[i] = NodePtr->Raw();
+				}
+			}
+		}
+
+		hashes[NumNodes] = Value;
+		return hashes;
+	}
 
 private:
 
-    bool NodeHasValue;
-    const uint8_t NumNodes = 16;
+	bool NodeHasValue;
+	const uint8_t NumNodes = 16;
 
-    std::vector<std::unique_ptr<TrieNode::NodeBase>> Branches;
-    SimpleObjects::Bytes Value;
+	std::vector<std::unique_ptr<Node>> Branches;
+	SimpleObjects::Bytes Value;
 }; // class BranchNode
 
 } // namespace Branch
