+++ test/src/TestEthEmptyNode.cpp
@@ -11,22 +11,23 @@
 
 using namespace EclipseMonitor;
 using namespace SimpleRlp;
+using namespace Trie;
 
 GTEST_TEST(TestEmptyNode, EmptyRawTest)
 {
-    std::vector<uint8_t> rlp = WriteRlp(TrieNode::EmptyNode::EmptyNodeRaw());
+    std::vector<uint8_t> rlp = WriteRlp(EmptyNode::EmptyNodeRaw());
     std::vector<uint8_t> expected = {0x80};
     EXPECT_EQ(rlp, expected);
 }
 
 GTEST_TEST(TestEmptyNode, EmptyNodeTest)
 {
-    std::vector<uint8_t> emptyNodeRlp = WriteRlp(TrieNode::EmptyNode::EmptyNodeRaw());
+    std::vector<uint8_t> emptyNodeRlp = WriteRlp(EmptyNode::EmptyNodeRaw());
     std::array<uint8_t, 32> emptyNodeHashArr = EthKeccak256(emptyNodeRlp);
 
     std::vector<uint8_t> emptyNodeHash;
     emptyNodeHash.reserve(32);
     std::move(emptyNodeHashArr.begin(), emptyNodeHashArr.end(), std::back_inserter(emptyNodeHash));
 
-    EXPECT_EQ(TrieNode::EmptyNode::EmptyNodeHash().GetVal(), emptyNodeHash);
+    EXPECT_EQ(EmptyNode::EmptyNodeHash().GetVal(), emptyNodeHash);
 }
